+++ src/main/kotlin/backend/itracker/tracker/controller/handler/ProductHandlerImpl.kt
@@ -1,7 +1,6 @@
-package backend.itracker.tracker.service
+package backend.itracker.tracker.controller.handler
 
 import backend.itracker.crawl.common.ProductCategory
-import backend.itracker.tracker.service.handler.ProductHandler
 import backend.itracker.tracker.service.response.filter.CommonFilterModel
 import backend.itracker.tracker.service.response.product.CommonProductDetailModel
 import backend.itracker.tracker.service.response.product.CommonProductModel
@@ -9,16 +8,13 @@ import backend.itracker.tracker.service.vo.Limit
 import backend.itracker.tracker.service.vo.ProductFilter
 import org.springframework.data.domain.Page
 import org.springframework.data.domain.PageRequest
-import org.springframework.stereotype.Service
-import org.springframework.transaction.annotation.Transactional
+import org.springframework.stereotype.Component
 
-@Transactional
-@Service
-class ProductService(
-    private val productHandlers: List<ProductHandler>
+@Component
+class ProductHandlerImpl(
+    private val productHandlers: List<ProductHandleable>
 ) {
 
-    @Transactional(readOnly = true)
     fun findTopDiscountPercentageProducts(
         productCategory: ProductCategory,
         limit: Limit
@@ -29,7 +25,6 @@ class ProductService(
         return productHandler.findTopDiscountPercentageProducts(productCategory, limit.value)
     }
 
-    @Transactional(readOnly = true)
     fun findFilter(
         productCategory: ProductCategory,
         productFilter: ProductFilter,
@@ -40,7 +35,6 @@ class ProductService(
         return productHandler.findFilter(productCategory, productFilter)
     }
 
-    @Transactional(readOnly = true)
     fun findFilteredProducts(
         category: ProductCategory,
         productFilter: ProductFilter,
@@ -52,7 +46,6 @@ class ProductService(
         return productHandler.findFilteredProductsOrderByDiscountRate(category, productFilter, pageable)
     }
 
-    @Transactional(readOnly = true)
     fun findProductById(
         category: ProductCategory, productId: Long
     ): CommonProductDetailModel {
